Add makemessages_fun translation commands
Added a makemessages_fun management command to collect translations for
fun-apps and the fun theme. Go to edxapp and run:

    ./manage.py lms makemessages_fun --settings=fun.lms_dev -l fr all

This should update the locale/fr/LC_MESSAGES/django.po file of all
fun-apps and the fun theme by:

    - adding new, required translations
    - keeping translations that override edx translations
    - removing obsolete translations

Also, update the various django.po files with this script.

Previously, fun theme was not translated at all, as it was all in
French. Now that we handle theme translations (via makemessages_fun), we
are able to localize the theme.

Note that translations are no longer collected in fun-apps/locale/, but
instead they are spread across the various applications, including
fun-apps/fun. Any application can thus override the edx translations.