makemessage_fun can be run as standalone script
We need to run makemessages_fun as a standalone script because pytz is
broken with distutils. Thus, in order to run makemessages_fun (or just
any pybabel script) you need a virtualenv with setuptools. Because the
edx virtualenv does not come with setuptools, we need to someone design
a workaround.