Use a file-based cache for ora2 file uploads
Unfortunately, we may not use memcache-based cache for ora2 file
uploads, because the cache needs to be shared by all web servers (and
FUN's memcaches are not shared). Thus, the only choices left are using a
file-based cache or a database. The database cache is a little painful
to maintain, as we need to run one "createcachetable" command per table.
Since there will be very few ora2 file submissions (an estimated less
than 1/s), we may safely use a slow, file-based cache.