Now using a custom class for staticfiles BowerFinder.
Why are we using a custom BowerFinder class? The finder class provided
by django-bower uses the standard `django.contrib.staticfiles`.
There is an issue with that, because edX expects a finder class that
extends the external `staticfiles` module - not the standard
module form `contrib`. The custom finder class we are using here
is a copy of bower's finder that extends the external `staticfiles`
module.