Now using FUN's fork for django-bower.
Why are we using a fork of django-bower?
The finder class provided by the original django-bower uses the standard
`django.contrib.staticfiles`. There is an issue with that, because edX
expects a finder class that extends the external staticfiles module and
not the standard module form contrib. The forked version brings  finder
class that extends the external staticfiles module.