course filtering: initial courses are properly filtered
We need to take into account the url parameters in the course filtering.
This allows users to access a specific query via its url. Here, we add
the "rpp" and "page" parameters. Also, we make sure that url parameters
are taken into account when accessing the page for the first time.

This closes issue #1969.