Fix video teaser in course about page
There were two issues with the former way of doing:
- Flash-based videos are regularly blocked by firefox for security reasons.
- The Dailymotion postMessage API did not allow us to detect player
  events, so trying to play the video too early resulted in javascript errors.

Both these issues prevented video playing in Firefox. To circumvent this
problem, we get rid of flash format and we load the video in a different
way: the modal is filled with the iframe html code every time the modal
is shown. The iframe is created with an autoplay parameter so the videos
starts playing right after the video is loaded.

This closes issue #2178.