Remove reference to djpyfs
django-pyfs is a module that adds S3-like API to the local filesystem.
It was mistakenly added to our dev installed apps as a Django debug
toolbar app, but it is in fact used nowhere. Proof of that is that the
syncdb command fails because of djpyfs, thus SQL tables are never
created for this app. To reproduce the bug, run the following command
before and after applying this patch:

        fun lms.dev syncdb # fails before, succeeds after.