Reorganize courses applications
The app called "courses" is the core application containing models and
utils. It's used by both the Studio and the LMS. This is also where we
trigger the updates of course meta data.
The app called "course_pages" is meant to be used on the LMS. It renders
front-end pages.
The app called "courses_api" only contains API related things.

* Move filters in courses api module
* Update translations
* Add courses pages app and move pages related elements there
* Change courses API URL and base name