Remove contact form and invite user to send an email.
For quality and security Mailjet service requests validation of email sender before allowing traffic, but we need technical support emails to have requester address as sender for correct handling by Zendesk.
While the problem should be fixed by configuring Postfix to not use Mailjet for emails sent to contact@france-universit-numerique-mooc.fr, this commit is a quickfix to stop losing user requests.

This commit should be reversed in a while.