Fix video teaser issue in Firefox
Video teaser does not show up in Firefox. As investigated by
univ-lorraine, this is due to the fact that the video modal dialog
appears with a "transform: translation(i0px, 0px)" css rule. Solution
was found here: http://stackoverflow.com/questions/20796646/video-not-working-in-bootstrap-modal-pop-up-on-firefox
(Note that we don't have the true explanation, though)

This closes issue #2178.