Fix news article created_at date in backoffice form
The created_at field needs to have a formatted recognized by the
javascript datetime parsing library. This format needs to be the same
for all languages, in particular when the logged-in user has selected
the French language. Thus, we need to specify the behaviour of both the
form field and the form widget.

This addresses (partially) issue #2316.