Cache libcast calls
Here, we try to reduce the number of calls made synchronously to the
Libcast API. To do so, we store responses from the API in a file-based
cache. Whenever a resource is updated, its cache is deleted. Calls from
xblocks should mostly land in the cache.

This PR comes with a joint PR from libcast-xblock.

This closes issue #2411.