Fix broken teaser when flash is deactivated
In a browser where flash is deactivated, the dailymotion video of the
course teaser does not play. This is due to the fact that the browser
does not correctly recognize that flash was deactivated. Previously, the
"&html" parameter was recognized by dailymotion, but it's not recognized
anymore. Instead, we need to use the "&html=1" parameter to make sure we
fetch the html5 video.

This should close issue #2515.