Fix answer parsing in course dashboard problem stats
When a problem is changed from a multiple choice question (with multiple
valid answers but only one answer per student) to a checkbox problem
(with multiple possible answers per student) the quizz answers failed to
be correctly evaluated.

Here, we decide that single answers to checkbox problems are valid. We
also fix a parsing issue for problems with more than 10 possible
answers. Finally, we also evaluate invalid answers to blank.

This closes issue #2508.