Do not translate LANGUAGES in the settings
Since the settings module is executed at the moment the server is run,
and not once per user, it makes no sense to translate the platform
languages in the settings. Instead, the language of each entry in the
"LANGUAGES" variable should appear in the original language of each
entry.

This partially addresses issue #2559.