Change i18n command to remove line numbers from .po files
Line numbers that change are annoying when making pull requests. They
increase the size of diffs uselessly. In this commit we remove the line
numbers from the i18n files but manage to keep the file path. To achieve
that, we have had to replace line numbers by "0", because message
formatting requires an integer as string in pybabel.