Make homepage template SQL requests run at every display
Mako code with <%! %> is executed only once, when the template is loaded
in memory. Thus, if we need to re-execute an SQL query for every page
load, we need to place the queries within <% %> tags.

Note that this will significantly increase the number of SQL requests.

This partially addresses issue #2537.