Fix certificate download from backoffice in development mode
In development mode, the /static/attestations/... produces a 404. That
is because we are not using Nginx to route such requests to the
attestations/ folder. In order to use these urls in the backoffice, we
define an /attestations/... url and use it in the backoffice, pretty
much like it is used in the student dashboard.