Fix university name on test certificate
Test certificates could in some cases have a different university name
than the one that is indicated on real certificates. This is because the
get_certificate_params was used in one case, and not the other. Here,
we make sure that the first university name is used in both cases.