Display registered users information from proctorU API in backoffice
A request is made each time we load the page, so there are some
performances issues.

We can't filter by courses when we query the API, so we do it server
side. And we can't ask for a specific test, so we need to give huge time
ranges.

This closes #2841