Fix uploads of large videos to libcast
This fixes CORS errors that were observed when uploading videos to
libcast. These errors were occurring for instance when uploading
multiple videos concurrently. This is because some headers are set
globally by edx-platform for all ajax requests, including OPTIONS
requests made for CORS. To avoid this problem, we were unsetting the
headers prior to the upload. But the headers were being reset before the
second chunk of large files started uploading. This was causing the
addition of unwanted headers to the CORS request for the second (and
third, etc.) chunk.

To solve this problem, we make sure the headers are unset just before
every chunk upload.

This closes issue #2918.