Increase libcast timeout for account creation
For a course that was just created, we need to list the files at the
root of the account. The problem is that this call can exceed the 10s
tolerated timeout for accounts that hold a great number of videos (such
as MinesTelecom). To circumvent this issue, which prevents universities
to upload videos right after they created the course, we increase the
allowed timeout of the "files/" endpoint.