Generate FUN verified certificates
We create a celery task to generate FUN verified certificates. The list
of users who deserve a certificate are also listed in the backoffice
("Verified Users" tab). We also introduce the
generate_fun_verified_certificates management command to generate
certificates manually. For the celery task or the management command, we
may ignore the proctoru reports by defining the PROCTORU_ALL_STUDENTS_OK
environment variable.  This is useful for debug.

Note that the backoffice page remains quite slow, but this is acceptable
for ~300 verified users.

Remains to be done: buttons in the backoffice to generate verified
certificates (issue #1836).

This closes issues #2833.