Verify T&C compliance on student dashboard page only
We need to verify the compliance of T&Cs for all connected users.
Previously, this was done for all users, on all pages and resulted in a
great number of http requests (one for every user and every page) and
sql requests (3 for every connected users, including on the
CourseEnrollment table which does not have an index for the 'mode'
column). To decrease the load, we check the compliance of payment terms
T&C on the dashboard page only, which is displayed to logged-in users
only.