Fix infinite redirect in backoffice for regular users
Non-admin users who were trying to access the backoffice were facing
infinite redirects between the login page and the backoffice page. This
is because we should not redirect users to the login page when they are
already logged-in.

In this commit, we return a 404 error in case a logged-in user tries to
access a backoffice page. Technically, we should display a 403 error,
but I believe they are not properly handled by the application (there is
no 403.html template page). So we return a 404 error instead.

This partially addresses issue #2851.