Deduplicate has_verified_course_mode information from SQL
has_verified_course_mode was being added by the update_courses script
while the information was already present in the CourseMode table. Here,
we deduplicate this information by using a cache that is local to the
fun Course model. This cache is only valid for a couple seconds and
should thus expire after every couple requests.