Integrate videofront as a videoprovider
Add videofront as a videoprovider in fun-apps.
Note that the video upload dashboard still uses youtube, so youtube
remains the default video provider. This is useful for testing the
compatibility of the video xblock with the videofront provider without
affecting the dashboard currently in place.

Note that we do not support thumbnail upload, yet.