fix(teaser): don't show video button when video missing
The creator of a course can link a teaser in Studio.
This saves a youtube url in Mongodb. This behavior was
hacked to allow creators to enter dailymotion IDs and
rebuild Dailymotion urls from the Youtube url. The hack
backfires when the user does not enter any ID in Studio.
This quick fix aims to hide the video button in this case
and avoid building a DM url pointing to nothing.