feat(search): improve ordering of search results
We introduce sorting by a variable field. This is because
when a course has not started yet, the start date is shown
and interesting, whereas once the course has started, the
end of enrollment date becomes more relevant and is shown.
The date used for sorting is now the date that is shown for
each course.