🐛(legal) stop redirecting impersonating staff to legal acceptance
When a staff user was impersonating a user who has not yet accepted
the next version of legal terms and conditions, he was redirected
to the terms and conditions acceptance page. We don't want the staff
user to accept the T&C in place of the user and we want to be able
to continue inmpersonating a user who is in this case.