🐛(legal) stop redirecting impersonating staff to legal acceptance
The problem is that when a staff user is impersonating a user who
has not yet accepted the latest version of our legal terms and
conditions, he is redirected to the T&C acceptance page. We don't
want the staff user to accept the T&C in place of the user and we
want to be able to continue impersonating a user who is in this
case so let's disable the redirect in this case.