⚡️(funsite) cache home page for authenticated users except staff
The home page was cached only for anonymous users and was very slow
for authenticated users (>5 seconds). The content of the home page
is the same for everybody, except for the header, so we can cache all
its body for everybody except staff users who will want to see their
modifications appear immediately.