⚡️(funsite) accelerate home page by overriding edX's view
In the previous release, we cached the body of the home page template
for authenticated users in an attempt to speed it up and found out that
most of the time is spent in edX's view where they retrieve objects that
we don't use in our home page. Adding our own view should further
accelerate our home page.