✨(payment) after payment automatically add verified users to a cohort
When users complete a payment make sure that they are added to a cohort.
The cohort name can be defined per course, in the
settings.VERIFIED_COHORTS list. This list includes (r"<course id
regex>", "<cohort name>") tuples. Users are added to the last cohort
from this list for which the course ID matches the regular expression.