✨(videoproviders) try loading VideoFront videos from metadata file
We want to protect against VideoFront server downtimes. A ready-made
VideoFront video document was added as a metadata json file in the S3
bucket alongside each video.

We first try getting the video document from this metadata json file,
and fall back to the previous behavior: requesting it to the VideoFront
API.