🐛(fun) fix conditional admin urls
The registration of the admin interface for the universities app was
limited to the LMS app. The utils checking this was based on the path
of the urls module but we changed its name to further extend it.
There is a better way to check this with the `SITE_VARIANT` setting.