✨(mailing_list) create views to un.subscribe to course mail
The learner dashboard will be soon replaced. To un.subscribe to a
course's mailing list, the user can do that only from the dashboard
course list view. As this view will be hidden, we need to find a
workaround. So the workaround is to create a public route to un
.subscribe a user to a course's mailing list. In this way, we will be
able to put a unsubcribe link into course's emails.