✨(courses) add title to course run synchronization data
When the course code of a course run (in the course key) is
not known by the remote site, we want to be able to automatically
create it. For this, it would be nice to set a title for the
course we will create. In MongoDB there is a "display_name" field
with this info.