✨(fun) add middleware to inspect oauth2 authorize request
When Moodle initiate an Oauth2 request, the state variable in the querystring
contains itself a querystring. On the redirect on the login page when a user is
not connected, Django set in a next variable the initial request path and then redirect on this
request path once connected. But the state is unquoted and splited in multiple variable in the querystring...
To fix this, first we store the initial query string in the request object and then we rebuild
the next query string by quoting once again the state to ensure it will not be splited
once redirected.